+++ packages/native/storage/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - fix!: Don't set a default namespace on Web
 - fix!: Align namespace/scope prefix with `shared_preferences` on Web
+- fix: Return deleted value
 
 ## 0.2.3